requirements.txt
ultralytics
opencv-python
streamlit
pillow
